fixed multiple login issues
