+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 
-import 'package:project_eureka_flutter/login_page.dart';
+import 'file:///C:/Users/Tony/Desktop/PE/project-eureka-flutter/lib/screens/login_page.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
