+++ lib/screens/forgot_password.dart
@@ -1,0 +1,53 @@
+import 'package:flutter/material.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+
+class ForgotPassword extends StatefulWidget {
+  @override
+  _ForgotpasswordState createState() => _ForgotpasswordState();
+}
+
+class _ForgotpasswordState extends State<ForgotPassword> {
+  String _email;
+  final auth = FirebaseAuth.instance;
+  @override
+  void initState() {
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(title: Text('Reset Password')),
+        body: Container(
+          color: Colors.lightBlueAccent[100],
+          child: Column(
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: <Widget>[
+                TextFormField(
+                  keyboardType: TextInputType.emailAddress,
+                  decoration: const InputDecoration(
+                    hintText: 'Email',
+                  ),
+                  onChanged: (value) {
+                    setState(() {
+                      _email = value;
+                    });
+                  },
+                ),
+                Padding(
+                  padding: const EdgeInsets.symmetric(vertical: 16.0),
+                  child: ElevatedButton(
+                    onPressed: () {
+                      auth.sendPasswordResetEmail(email: _email);
+                      SnackBar(
+                        content: const Text('request sent'),
+                      );
+                      Navigator.of(context).pop();
+                    },
+                    child: Text('Send Request'),
+                  ),
+                ),
+              ]),
+        ));
+  }
+}
