+++ pubspec.yaml
@@ -31,6 +31,8 @@ dependencies:
   firebase_core: ^0.5.0
   firebase_auth: ^0.18.0+1
   google_sign_in: ^4.5.3
+  modal_progress_hud: ^0.1.3
+
 
 dev_dependencies:
   flutter_test:
